requirements.txt
firebase-admin==6.6.0
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
psycopg2==2.9.10
python-dotenv==1.0.1
reportlab==4.3.1
requests==2.32.3
scikit-learn==1.6.1
streamlit==1.41.1
streamlit-option-menu==0.4.0
supabase==2.11.0
matplotlib==3.9.2
